matrix

+++ app/src/main/java/com/example/myapplication3/Entity.java
@@ -4,7 +4,7 @@ import android.graphics.Bitmap;
 
 public class Entity {
     Bitmap bitmap;
-    int x, y, width, height;
+    float x, y, width, height;
 
     public Entity(Bitmap bitmap, int x, int y, int width, int height) {
         this.bitmap = bitmap;
@@ -22,35 +22,35 @@ public class Entity {
         this.bitmap = bitmap;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 
-    public int getWidth() {
+    public float getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(float width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public float getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(float height) {
         this.height = height;
     }
 }
